+++ test/fixtures/synthetic/escape-hatch-book/styles/site.clj
@@ -1,2 +1,0 @@
-[:body {:color "rebeccapurple"
-        :font-style "italic"}]